requirements.txt
timm>=0.9.12
opencv-python>=4.5.0          
Pillow>=9.0.0                 
numpy>=1.22.0
scikit-learn>=1.0.0           
matplotlib>=3.5.0             
streamlit>=1.28.0
tqdm>=4.60.0                  